reset: Bringing up to date with ‘klanger 2.0’ which is now live
This is the from-scratch version of klanger 2.0, formerly ‘klaxon’,
which is now live on the site.  It’s still not finished, but it’s
90% of the way there.  It only needs some finishing touches and
quality of life improvements, mainly around share-sheets.

I’ve also included easter-eggs.js as a separate file for ease of
updating and maintainance; I might later see if I can have Tumblr
serve this separately rather than inline.  (Same, perahps, for the
CSS file.)

Since this is a reset, I’ve deleted the old theme.html version.